+++ src/version.d.ts
@@ -1,3 +1,3 @@
-export declare const SDK_VERSION = "2.0.2";
-export declare const PROTOCOL_VERSION = "2.0";
+export declare const SDK_VERSION = "3.0.0";
+export declare const PROTOCOL_VERSION = "3.0";
 export declare const SUPPORTED_PROTOCOL_VERSIONS: readonly string[];